step2 파일에서 코드 수정하기
기존 step3 과제 파일을 제거하고 step2 파일에 이어서 진행
