--- Task1/Assets/Scripts/DicomDataSearch.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 99275966b0f49ee47be8464e06578dc9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
